Moved liquibase.Activator to liquibase.osgi.Activator
