+++ liquibase-core/pom.xml
@@ -159,7 +159,7 @@
                 <configuration>
                     <instructions>
                         <Bundle-SymbolicName>org.liquibase.core</Bundle-SymbolicName>
-                        <Bundle-Activator>liquibase.Activator</Bundle-Activator>
+                        <Bundle-Activator>liquibase.osgi.Activator</Bundle-Activator>
                         <Import-Package>
                             javax.activation*;resolution:=optional,
                             javax.servlet.*;version="[2.6,4)";resolution:=optional,
